BT02

+++ BTVN/BT02/BT02C3.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 
 int main() {
-    long long x;
+    int x;
     cin >> x;
     for(int i = 1; i <= x; ++i) {
         for(int j = i; j < x; ++j) {
@@ -15,3 +15,4 @@ int main() {
         cout << endl;
     }
 }
+
