+++ BTVN/BT02/BT02C4.cpp
@@ -1,9 +1,8 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-
-int main() {
-    long long x;
+int main(){
+    int x;
     cin >> x;
     for(int i = x; i >= 1; --i) {
         for(int j = x; j > i; --j) {
